Improved upstart scripts
The depencency of start/stop was not declared on the rasp-telegram
script, so it was not being restarted automatically if the
telegram service was restarted.
The start/stop conf was set on the conf file. So, now if the
telegram-daemon is started or stoped the rasp-telegram is also
started/stoped.